requirements/pip_requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloud-init==22.1
colorama==0.4.3
commentjson==0.9.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.11.0
Cython==0.29.28
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
fonttools==4.25.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.3
ImageHash==4.2.1
imageio==2.16.1
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keyring==18.0.1
kiwisolver==1.3.2
lark-parser==0.7.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==4.2.0
munkres==1.1.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.8
numpy==1.22.4
oauthlib==3.1.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.7.1
pycparser==2.21
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.27.1
requests-unixsocket==0.2.0
scipy==1.7.3
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sos==4.3
soupsieve==2.3.2.post1
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
terminado==0.13.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
Twisted==18.9.0
typing_extensions==4.1.1
ufw==0.36
urllib3==1.26.9
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
wheel>=0.37.1
widgetsnbextension==3.5.2
yarg==0.1.9
zipp==1.0.0
zope.interface==4.7.1